% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimateCortisol.R
\name{backEstimateCortisol}
\alias{backEstimateCortisol}
\title{Back-estimating Cortisol Concentration}
\usage{
backEstimateCortisol(dat, logStdVar, logValVar, calibVar)
}
\arguments{
\item{dat}{Should be the data saved from the 'initialInspectionCortisol' function}

\item{logStdVar}{The logStdVar variable from the data.}

\item{logValVar}{The logStdVar variable from the data.}

\item{calibVar}{The calibVar variable from the data.}
}
\description{
Back-estimating Cortisol Concentration
}
\examples{

}

