% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInspectionCortisol.R
\name{initialInspectCortisol}
\alias{initialInspectCortisol}
\title{Initial Data Inspection}
\usage{
initialInspectCortisol(dat, calibName, calibValue, nVar, drop)
}
\arguments{
\item{dat}{A data frame}

\item{calibVar}{The name of the calibration variable (usually STAND or A)}

\item{calibValue}{The value for calibration (usually 0)}

\item{cpmVar}{The name for the CPM variable (usually CPM)}

\item{idVar}{The ID variable; is often "N".}

\item{drop}{Which participant to drop.}
}
\description{
This function does an intial inspection on the data.
}
\examples{

dat = data.frame(N = 1:10,
                 CPM = c(44396.5, 47774.5, 23676.0, 24290.3, 23541.0,
                         20108.7, 20101.3, 19383.7, 17013.7, 17678.7),
                 STAND = c(NA, NA, 0, 0, 1,
                           1, 3, 10, 30, 100))
# The First Pass #

initialInspectCortisol(dat, "STAND", 0, "CPM")

# Deleting Observations #

initialInspectCortisol(dat, "STAND", 0, "CPM", "N", 1)

}

