% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple ggplot2 Objects}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Any number of ggplot2 objects.}
}
\description{
This function takes a named ggplot2 objects and plots them in a single view.
}
\examples{
multiplot()
}
\author{
Michael Clark
}

