% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Determine Lower Detection Threshold}
\usage{
sensitivity(dat, calibVar, calibVal, logStdVar, logValVar)
}
\arguments{
\item{dat}{The data object returned from initialInspectionCortisol}

\item{logStdVar}{The logStd variable from the data}

\item{logValVar}{The logVal variable from the data}
}
\description{
Determine Lower Detection Threshold
}

