% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimateCortisol.R
\name{backEstimateCortisol}
\alias{backEstimateCortisol}
\title{Back-estimating Cortisol Concentration}
\usage{
backEstimateCortisol(dat, calibVar)
}
\arguments{
\item{dat}{Should be the data saved from the 'initialInspectionCortisol' function}

\item{calibVar}{The calibVar variable from the data.}
}
\description{
Back-estimating Cortisol Concentration
}
\examples{
dat = data.frame(N = 1:10,
                 CPM = c(44396.5, 47774.5, 23676.0, 24290.3, 23541.0,
                         20108.7, 20101.3, 19383.7, 17013.7, 17678.7),
                 STAND = c(NA, NA, 0, 0, 1,
                           1, 3, 10, 30, 100),
                 bindings = c(445.7136, 479.6267, 237.6925, 243.8597,
                              236.3372, 201.8790, 201.8047, 194.6005,
                              170.8071, 177.4833),
                 logVals = c(-1.13, -1.34, -1.18, 0.1, .08,
                             .32, -1.47, -1.65, -1.61, -.74),
                 logStd = c(0, 0, 0, 1.09, 1.09,
                            1.1, 2.30, 2.30, 2.30, 3.40))
backEstimateCortisol(dat, "STAND")
}

